requirements.txt
flask
torch
sentence-transformers
transformers
python-dotenv
pymupdf
pandas
uuid
nltk